refactor: move BaseIdEntity to support.domain.entity
