+++ src/main/java/com/meoguri/linkocean/domain/tag/entity/Tag.java
@@ -7,7 +7,7 @@ import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Table;
 
-import com.meoguri.linkocean.domain.BaseIdEntity;
+import com.meoguri.linkocean.support.domain.entity.BaseIdEntity;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
